# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  version: '1.0'
  description: |-
    API per l'applicazione WASAText che consente di chattare con le persone o gruppi di persone.
    Gli endpoint che agiscono sull'utente autenticato (es. setMyUserName, setMyPhoto)
    non richiedono un User ID nel percorso, poiché l'identificativo viene estratto
    dall'header di autorizzazione della richiesta.
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
...
paths:
...
  /me/name:
    put:
      tags: ["user"]
      summary: Updates the name of the authenticated user
      description: The name must be unique and adheres to the specified pattern.
      operationId: setMyUserName
      requestBody:
        description: New name for the authenticated user
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  $ref: "#/components/schemas/Name"
              required:
                - name
      responses:
        "200":
          description: Name changed successfully
          content: {}
        "409":
          description: New name is already in use
        "401":
          description: Unauthorized
  /me/photo:
    put:
      tags: ["user"]
      summary: Updates the authenticated user's profile photo
      description: Uploads a new profile photo for the authenticated user.
      operationId: setMyPhoto
      requestBody:
        description: New profile photo file
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: New profile image file (e.g., JPEG, PNG)
              required:
                - image
      responses:
        "200":
          description: Photo changed successfully
          content: {}
        "400":
          description: Invalid input, missing file or incorrect format
        "401":
          description: Unauthorized
  /users/search:
    get:
      tags: ["users"]
      summary: Searches for a user by name
      description: Returns a list of users whose names match the search query.
      operationId: searchUsers
      parameters:
        - in: query
          name: name
          schema:
            type: string
            minLength: 1
            maxLength: 50
          required: true
          description: The username substring to search for.
      responses:
        "200":
          description: A list of users found
          content:
            application/json:
              schema:
                type: array
                minItems: 0
                maxItems: 100 # Aggiunto maxItems per l'array
                items:
                  $ref: "#/components/schemas/User"
...
  /messages/{messageId}:
    parameters:
      - $ref: "#/components/parameters/messageId"
    delete:
      tags: ["messages"]
      summary: Deletes a sent message
      description: Allows the sender to delete a message they have sent.
      operationId: deleteMessage
      responses:
        "200":
          description: Message deleted successfully
          content: {}
        "403":
          description: User is not the sender of the message
        "404":
          description: Message not found
        "401":
          description: Unauthorized
  /messages/{messageId}/forward:
    parameters:
      - $ref: "#/components/parameters/messageId"
    post:
      tags: ["messages"]
      summary: Forwards an existing message to multiple conversations
      description: Sends a copy of the message to a list of other conversations.
      operationId: forwardMessage
      requestBody:
        description: List of conversation IDs to forward the message to
        content:
          application/json:
            schema:
              type: object
              properties:
                conversationIds:
                  type: array
                  items:
                    $ref: "#/components/schemas/Id"
                  minItems: 1
                  maxItems: 50
                  description: List of conversation IDs to forward the message to.
              required:
                - conversationIds
      responses:
        "200":
          description: Message forwarded successfully
          content: {}
        "403":
          description: User is not authorized to forward to one or more conversations
        "404":
          description: Message or conversation not found
        "401":
          description: Unauthorized
  /messages/{messageId}/reactions:
    parameters:
      - $ref: "#/components/parameters/messageId"
    post:
      tags: ["messages"]
      summary: Adds a reaction to a message
      description: Adds an emoji reaction to the specified message.
      operationId: commentMessage
      requestBody:
        description: Emoji for the reaction
        content:
          application/json:
            schema:
              type: object
              properties:
                emoji:
                  $ref: "#/components/schemas/Reaction"
              required:
                - emoji
      responses:
        "200":
          description: Reaction added successfully
          content: {}
        "404":
          description: Message not found
        "401":
          description: Unauthorized
    delete:
      tags: ["messages"]
      summary: Removes a reaction from a message
      description: Removes the authenticated user's reaction to a message.
      operationId: uncommentMessage
      responses:
        "200":
          description: Reaction removed successfully
          content: {}
        "404":
          description: Reaction or message not found
        "401":
          description: Unauthorized
...
components:
  schemas:
    Id:
      description: A unique identifier (integer).
      type: integer
      example: 1234
    Name:
      description: A new username or group name, between 3 and 16 characters.
      type: string
      example: "Emanuele Panizzi"
      minLength: 3
      maxLength: 16
    User:
      type: object
      description: Represents a user with basic profile information.
      properties:
        id:
          $ref: "#/components/schemas/Id"
        name:
          $ref: "#/components/schemas/Name"
        photo:
          $ref: "#/components/schemas/PhotoUrl"
      required:
        - id
        - name
        - photo # ERRORE CORRETTO: Aggiunto 'required' per User
    PhotoUrl:
      description: URL to a photo resource.
      type: string
      example: "/user/profile-photo/123"
      minLength: 1
      maxLength: 255 # Aggiunto maxLength
    Reaction:
      description: Emoji used in a reaction (single character).
      type: string
      example: "👍"
      minLength: 1
      maxLength: 1
...
  parameters:
...
    messageId:
      name: messageId
      in: path
      required: true
      description: The unique identifier for a message.
      schema:
        $ref: "#/components/schemas/Id"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT